rearranged tutorial steps
