--- examples/tutorial/step2/models.py
@@ -1,28 +1,0 @@
-import json
-from dataclasses import dataclass
-from typing import Optional, TypeVar, Type
-
-from rsocket.frame_helpers import ensure_bytes
-from rsocket.helpers import utf8_decode
-from rsocket.payload import Payload
-
-
-@dataclass(frozen=True)
-class Message:
-    user: Optional[str] = None
-    content: Optional[str] = None
-
-
-def encode_dataclass(obj):
-    return ensure_bytes(json.dumps(obj.__dict__))
-
-
-def dataclass_to_payload(obj) -> Payload:
-    return Payload(encode_dataclass(obj))
-
-
-T = TypeVar('T')
-
-
-def decode_dataclass(payload: Payload, cls: Type[T]) -> T:
-    return cls(**json.loads(utf8_decode(payload.data)))
