+++ examples/tutorial/step3/chat_client.py
@@ -1,16 +1,15 @@
 import asyncio
 import json
 import logging
-from typing import List, Optional
+from typing import Optional
 
 from examples.tutorial.step3.models import Message, encode_dataclass
 from reactivestreams.subscriber import DefaultSubscriber
 from reactivestreams.subscription import DefaultSubscription
-from rsocket.awaitable.awaitable_rsocket import AwaitableRSocket
 from rsocket.extensions.helpers import composite, route
 from rsocket.extensions.mimetypes import WellKnownMimeTypes
 from rsocket.frame_helpers import ensure_bytes
-from rsocket.helpers import single_transport_provider, utf8_decode
+from rsocket.helpers import single_transport_provider
 from rsocket.payload import Payload
 from rsocket.rsocket_client import RSocketClient
 from rsocket.transports.tcp import TransportTCP
@@ -29,20 +28,10 @@ class ChatClient:
         self._session_id = (await self._rsocket.request_response(payload)).data
         return self
 
-    async def join(self, channel_name: str):
-        request = Payload(ensure_bytes(channel_name), composite(route('channel.join')))
-        await self._rsocket.request_response(request)
-        return self
-
-    async def leave(self, channel_name: str):
-        request = Payload(ensure_bytes(channel_name), composite(route('channel.leave')))
-        await self._rsocket.request_response(request)
-        return self
-
     def listen_for_messages(self):
         def print_message(data: bytes):
             message = Message(**json.loads(data))
-            print(f'{self._username}: from {message.user} ({message.channel}): {message.content}')
+            print(f'{self._username}: from {message.user}: {message.content}')
 
         class MessageListener(DefaultSubscriber, DefaultSubscription):
             def __init__(self):
@@ -77,16 +66,6 @@ class ChatClient:
         await self._rsocket.request_response(Payload(encode_dataclass(Message(username, content)),
                                                      composite(route('message'))))
 
-    async def channel_message(self, channel: str, content: str):
-        print(f'Sending {content} to channel {channel}')
-        await self._rsocket.request_response(Payload(encode_dataclass(Message(channel=channel, content=content)),
-                                                     composite(route('message'))))
-
-    async def list_channels(self) -> List[str]:
-        request = Payload(metadata=composite(route('channels')))
-        response = await AwaitableRSocket(self._rsocket).request_stream(request)
-        return list(map(lambda _: utf8_decode(_.data), response))
-
 
 async def main():
     connection1 = await asyncio.open_connection('localhost', 6565)
@@ -110,13 +89,7 @@ async def messaging_example(user1: ChatClient, user2: ChatClient):
     user1.listen_for_messages()
     user2.listen_for_messages()
 
-    await user1.join('channel1')
-    await user2.join('channel1')
-
-    print(f'Channels: {await user1.list_channels()}')
-
     await user1.private_message('user2', 'private message from user1')
-    await user1.channel_message('channel1', 'channel message from user1')
 
     await asyncio.sleep(1)
