+++ examples/tutorial/step3/models.py
@@ -1,8 +1,9 @@
 import json
 from dataclasses import dataclass
-from typing import Optional
+from typing import Optional, TypeVar, Type
 
 from rsocket.frame_helpers import ensure_bytes
+from rsocket.helpers import utf8_decode
 from rsocket.payload import Payload
 
 
@@ -10,10 +11,6 @@ from rsocket.payload import Payload
 class Message:
     user: Optional[str] = None
     content: Optional[str] = None
-    channel: Optional[str] = None
-
-
-chat_filename_mimetype = b'chat/file-name'
 
 
 def encode_dataclass(obj):
@@ -22,3 +19,10 @@ def encode_dataclass(obj):
 
 def dataclass_to_payload(obj) -> Payload:
     return Payload(encode_dataclass(obj))
+
+
+T = TypeVar('T')
+
+
+def decode_dataclass(payload: Payload, cls: Type[T]) -> T:
+    return cls(**json.loads(utf8_decode(payload.data)))
