+++ examples/tutorial/test_tutorials.py
@@ -11,11 +11,12 @@ import pytest
                          [
                              'step0',
                              'step1',
-                             'step1_1',
                              'step2',
                              'step3',
                              'step4',
                              'step5',
+                             'step6',
+                             'step7',
                              'reactivex']
 
                          )
